add userId to all items + firebase function onCreate user

+++ lib/views/ingredients.dart
@@ -3,10 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:pom/blocs/ingredient/ingredient.dart';
 import 'package:pom/blocs/ingredient/ingredient_events.dart';
+import 'package:pom/main.dart';
 import 'package:pom/models/ingredient.dart';
 import 'package:pom/theme/themes.dart';
 import 'package:pom/views/ingredient.dart';
 import 'package:pom/widgets/confirm_action_dialog.dart';
+import 'package:pom/widgets/custom_appbar.dart';
 import 'package:pom/widgets/item_card.dart';
 
 class IngredientsPage extends StatefulWidget {
@@ -22,19 +24,20 @@ class _IngredientsPageState extends State<IngredientsPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: const Text('Ingrédients'),
-      ),
+      appBar: const CustomAppBar(title: Text('Ingrédients')),
       body: StreamBuilder<QuerySnapshot<Object?>>(
         stream: FirebaseFirestore.instance
             .collection('ingredients')
             .orderBy('name')
+            .where('userId', isEqualTo: firebaseAuth.currentUser!.uid)
             .snapshots(),
         builder: (
           BuildContext context,
           AsyncSnapshot<QuerySnapshot<Object?>> snapshot,
         ) {
-          if (!snapshot.hasData) return const LinearProgressIndicator();
+          if (!snapshot.hasData) {
+            return const Center(child: CircularProgressIndicator());
+          }
           final List<Ingredient> ingredients = snapshot.data == null
               ? <Ingredient>[]
               : snapshot.data!.docs
@@ -61,7 +64,7 @@ class _IngredientsPageState extends State<IngredientsPage> {
                               return const ConfirmActionDialog();
                             },
                           );
-                          if (shouldDelete != null && shouldDelete) {
+                          if (shouldDelete != null && shouldDelete && mounted) {
                             context.read<IngredientBloc>().add(
                                   DeleteIngredientByIdEvent(ingredient),
                                 );
